backend tweaks

+++ backend/controllers/product.controller.js
@@ -6,7 +6,7 @@ export const getProducts = async (req, res) => {
     res.status(200).json({ success: true, data: products });
   } catch (error) {
     console.log("Error while fetching products", error.message);
-    res.status(500).json({ success: false, message: "Server error" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
 
@@ -25,7 +25,7 @@ export const createProduct = async (req, res) => {
     res.status(200).json({ success: true, data: updateProdct });
   } catch (error) {
     console.log("Error while updating product", error.message);
-    res.status(500).json({ success: false, message: "Server error" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
 
@@ -43,19 +43,24 @@ export const updateProdct = async (req, res) => {
     res.status(201).json({ success: true, data: newProduct });
   } catch (err) {
     console.error("Errro while creating product", err.message);
-    res.status(500).json({ success: false, message: "server errr" });
+    res.status(500).json({ success: false, message: "server Error" });
   }
 };
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
   console.log("id: ", id);
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Product not found" });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res
       .status(200)
       .json({ success: true, message: `Product is deleted: ${id}` });
   } catch (error) {
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
